daemon_kafka_producer/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
kafka-python
requests==2.28.2
